// src/shared/enums/events.ts
export enum Events {
    // CitizenFX Events
    gameEventTriggered = "gameEventTriggered",
    entityCreated = "entityCreated",
    entityCreating = "entityCreating",
    entityRemoved = "entityRemoved",
    resourceListRefreshed = "onResourceListRefresh",
    resourceStart = "onResourceStart",
    resourceStarting = "onResourceStarting",
    resourceStop = "onResourceStop",
    serverResourceStart = "onServerResourceStart",
    serverResourceStop = "onServerResourceStop",
    playerConnecting = "playerConnecting",
    playerEnteredScope = "playerEnteredScope",
    playerLeftScope = "playerLeftScope",
    playerJoining = "playerJoining",
    playerDropped = "playerDropped",
    playerSpawned = "playerSpawned",

    // Base Events
    playerDied = "baseevents:onPlayerDied",
    playerKilled = "baseevents:onPlayerKilled",
    playerWasted = "baseevents:onPlayerWasted",
    enteringVehicle = "baseevents:enteringVehicle",
    enteringAborted = "baseevents:enteringAborted",
    enteredVehicle = "baseevents:enteredVehicle",
    leftVehicle = "baseevents:leftVehicle",

    // PGN Menu Client Events
    syncPlayers = "pgn-menu-client:syncPlayers",
    setKidnapped = "pgn-menu-client:setKidnapped",
    showNotification = "pgn-menu-client:showNotification",
    recieveAdvert = "pgn-menu-client:recieveAdvert",
    startCuffing = "pgn-menu-client:startCuffing",
    getUncuffed = "pgn-menu-client:getUncuffed",
    setGrabbed = "pgn-menu-client:setGrabbed",
    getSeated = "pgn-menu-client:getSeated",
    getUnseated = "pgn-menu-client:getUnseated",
    getHospitalized = "pgn-menu-client:getHospitalized",
    updateProps = "pgn-menu-client:updateProps",
    startSpeedzone = "pgn-menu-client:startSpeedzone",
    removeSpeedzone = "pgn-menu-client:removeSpeedzone",
    getJailed = "pgn-menu-client:getJailed",

    // PGN Menu Server Events
    playerLoaded = "pgn-menu-server:playerLoaded",
    kidnapPlayer = "pgn-menu-server:kidnapPlayer",
    sendAdvert = "pgn-menu-server:sendAdvert",
    cuffPlayer = "pgn-menu-server:cuffPlayer",
    uncuffPlayer = "pgn-menu-server:uncuffPlayer",
    grabPlayer = "pgn-menu-server:grabPlayer",
    seatPlayer = "pgn-menu-server:seatPlayer",
    unseatPlayer = "pgn-menu-server:unseatPlayer",
    startHospitalize = "pgn-menu-server:startHospitalize",
    exitBed = "pgn-menu-server:exitBed",
    syncProps = "pgn-menu-server:syncProps",
    createSpeedzone = "pgn-menu-server:createSpeedzone",
    deleteSpeedzone = "pgn-menu-server:deleteSpeedzone",
    jailPlayer = "pgn-menu-server:jailPlayer"
}